Change how changelogs are generated and used
The logic for pulling out the first release is useful for download but kind of sucks for changelog. Frankly I could pull it out entirely but it's kind of nice to have the one static "must be set" value.

Anyways, s/oldReleases/releaseHistory/ and change it so that only release history is used for changelog, the Latest.Version won't have changelog populated.